Added floating-point precision to camera, changed makefile obj for Mac.

+++ src/camera.c
@@ -47,8 +47,8 @@
 void
 update_camera( entity_t *focus_point ) {
   if ( focus_point != NULL ) {
-    app.camera.x = ( ( int32_t ) focus_point->x + focus_point->w / 2 ) - ( app.SCREEN_WIDTH >> 1 );
-    app.camera.y = ( ( int32_t ) focus_point->y + focus_point->h / 2 ) - ( app.SCREEN_HEIGHT >> 1 );
+    app.camera.x = ( focus_point->x + focus_point->w / 2 ) - ( app.SCREEN_WIDTH >> 1 );
+    app.camera.y = ( focus_point->y + focus_point->h / 2 ) - ( app.SCREEN_HEIGHT >> 1 );
     app.camera.w = app.SCREEN_WIDTH;
     app.camera.h = app.SCREEN_HEIGHT;
